# api/openapi.yaml
openapi: "3.1.0"
info:
  title: DirtOS REST API
  version: "1"
  description: |
    Local REST API embedded in the DirtOS desktop application.
    Intended for plugin development and 3rd-party integrations (n8n, Home Assistant, scripts, etc.).

    **Base URL:** `http://127.0.0.1:7272`
    Override the port with the `DIRTOS_API_PORT` environment variable before launching DirtOS.

    All endpoints return JSON. Timestamps are ISO 8601 strings in UTC.
    Enum values are lowercase snake_case strings unless noted otherwise.
servers:
  - url: http://127.0.0.1:7272
    description: DirtOS local API (default port)
...
paths:
...
  /api/v1/species:
    get:
      tags: [Species]
      summary: List / search species
      operationId: listSpecies
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Free-text search against common name and scientific name
        - name: sun_requirement
          in: query
          schema:
            type: string
        - name: water_requirement
          in: query
          schema:
            type: string
        - name: growth_type
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of species
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Species"
    post:
      tags: [Species]
      summary: Create a species record
      operationId: createSpecies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSpecies"
      responses:
        "200":
          description: Created species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Species"
  /api/v1/species/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Species]
      summary: Get a species record
      operationId: getSpecies
      responses:
        "200":
          description: Species record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Species"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Species]
      summary: Update a species record
      operationId: updateSpecies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpecies"
      responses:
        "200":
          description: Updated species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Species"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Species]
      summary: Delete a species record
      operationId: deleteSpecies
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Plants
  # ---------------------------------------------------------------------------
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
      description: Maximum number of records to return
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of records to skip (for pagination)
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not found
  schemas:
...
    Species:
      type: object
      properties:
        id:
          type: integer
          format: int64
        common_name:
          type: string
        scientific_name:
          type: string
          nullable: true
        family:
          type: string
          nullable: true
        genus:
          type: string
          nullable: true
        growth_type:
          type: string
          nullable: true
        sun_requirement:
          type: string
          nullable: true
        water_requirement:
          type: string
          nullable: true
        soil_ph_min:
          type: number
          nullable: true
        soil_ph_max:
          type: number
          nullable: true
        spacing_cm:
          type: number
          nullable: true
        days_to_germination_min:
          type: integer
          nullable: true
        days_to_germination_max:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewSpecies:
      type: object
      required: [common_name]
      properties:
        common_name:
          type: string
        scientific_name:
          type: string
          nullable: true
        family:
          type: string
          nullable: true
        genus:
          type: string
          nullable: true
        growth_type:
          type: string
          nullable: true
        sun_requirement:
          type: string
          nullable: true
        water_requirement:
          type: string
          nullable: true
        soil_ph_min:
          type: number
          nullable: true
        soil_ph_max:
          type: number
          nullable: true
        spacing_cm:
          type: number
          nullable: true
        days_to_germination_min:
          type: integer
          nullable: true
        days_to_germination_max:
          type: integer
          nullable: true
    UpdateSpecies:
      type: object
      properties:
        common_name:
          type: string
          nullable: true
        scientific_name:
          type: string
          nullable: true
        family:
          type: string
          nullable: true
        genus:
          type: string
          nullable: true
        growth_type:
          type: string
          nullable: true
        sun_requirement:
          type: string
          nullable: true
        water_requirement:
          type: string
          nullable: true
        soil_ph_min:
          type: number
          nullable: true
        soil_ph_max:
          type: number
          nullable: true
        spacing_cm:
          type: number
          nullable: true
        days_to_germination_min:
          type: integer
          nullable: true
        days_to_germination_max:
          type: integer
          nullable: true

    # -------------------------------------------------------------------------
    # Plant
    # -------------------------------------------------------------------------
...